res made the following changes: - Removed the fake bool stuff. Compilers that are able to deal with   CS' usage of templates are deemed to have a 'bool' type. - Some doxygen documentation fixes. - Replaced the CS_ASSERTs with CS_ASSERT_MSGs in csutil/ref.h. - Removed the flattening from iEvent. Instead, iEvent now offers   an attribute iterator which can be used to serialize an event   "from the outside". Moved flattening code into a separate helper   class, csEventFlattener. - Removed ability to store multiple event attributes with the same   name. - Added ability to store iBase data in event attributes. - Renamed iEvent::Find() to iEvent::Retrieve() due changed return   value (and semantics change caused by that). - Fixed copy'n'paste error in csendian.h. The int64 variant of   convert_endian() would use little_endian_long() instead of   little_endian_longlong(), causing bogus results. - Fixed some 64-bit warnings in VFS.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@18542 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/filesys/vfs/vfs.h
@@ -37,7 +37,7 @@ class csFile : public iFile
 {
 protected:
   // Index into parent node RPath
-  int Index;
+  size_t Index;
   // File node
   VfsNode *Node;
   // Filename in VFS
@@ -48,7 +48,7 @@ protected:
   int Error;
 
   // The constructor for csFile
-  csFile (int Mode, VfsNode *ParentNode, int RIndex,
+  csFile (int Mode, VfsNode *ParentNode, size_t RIndex,
   	const char *NameSuffix);
 
 public:
